+++ src/platforms/os.rs
@@ -55,7 +55,7 @@ impl Os {
 }
 
 /// Represents an OS version. In case of Linux this also includes the distro and distro version.
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 #[allow(dead_code)]
 pub enum OsVersion {
     MacOs {