+++ src/platforms/target.rs
@@ -2,7 +2,7 @@
 use crate::platforms::{OsVersion, TargetArchitecture};
 
 /// Represents a target, which is its architecture and OS version combined.
-#[derive(Debug)]
+#[derive(Debug, PartialEq)]
 pub struct Target {
     pub architecture: TargetArchitecture,
     pub os: OsVersion,